+++ src/sharc-creative.js
@@ -34,7 +34,7 @@
  * </script>
  * ```
  *
- * @version 0.5.2
+ * @version 0.5.3
  */
 
 'use strict';